Remove top-level ClientSideRoutingContext which wasn't actually needed
Simplify the client_side_routing.js implementation

Move initial hydrate back into `useEventLoop`
