+++ reflex/.templates/web/pages/_app.js
@@ -1,8 +1,7 @@
 import { ChakraProvider, extendTheme } from "@chakra-ui/react";
 import { Global, css } from "@emotion/react";
 import theme from "/utils/theme";
-import { ClientSideRoutingProvider } from "utils/client_side_routing";
-import { clientStorage, initialState, StateContext, EventLoopContext } from "/utils/context.js";
+import { clientStorage, initialEvents, initialState, StateContext, EventLoopContext } from "/utils/context.js";
 import { useEventLoop } from "utils/state";
 
 import '../styles/tailwind.css'
@@ -18,6 +17,7 @@ const GlobalStyles = css`
 function EventLoopProvider({ children }) {
   const [state, Event, connectError] = useEventLoop(
     initialState,
+    initialEvents,
     clientStorage,
   )
   return (
@@ -34,9 +34,7 @@ function MyApp({ Component, pageProps }) {
     <ChakraProvider theme={extendTheme(theme)}>
       <Global styles={GlobalStyles} />
       <EventLoopProvider>
-        <ClientSideRoutingProvider>
-          <Component {...pageProps} />
-        </ClientSideRoutingProvider>
+        <Component {...pageProps} />
       </EventLoopProvider>
     </ChakraProvider>
   );
